some improvements and cleanup
