+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "vite-plugin-ziggy",
-  "version": "0.3.0",
+  "version": "0.2.0",
   "description": "Vite plugin for Ziggy",
   "keywords": [
     "vite",
@@ -9,12 +9,24 @@
   ],
   "type": "module",
   "exports": {
-    ".": {
-      "default": "./dist/index.js",
-      "types": "./dist/index.d.ts"
+    "./build": {
+      "default": "./dist/build.js",
+      "types": "./dist/build.d.ts"
+    },
+    "./config": {
+      "default": "./dist/config.js",
+      "types": "./dist/config.d.ts"
+    },
+    "./utils": {
+      "default": "./dist/utils.js",
+      "types": "./dist/utils.d.ts"
     },
     "./routes": {
       "types": "./routes.d.ts"
+    },
+    ".": {
+      "default": "./dist/index.js",
+      "types": "./dist/index.d.ts"
     }
   },
   "types": "./dist/index.d.ts",
