+++ src/index.ts
@@ -1,147 +1,19 @@
-import { readFileSync } from 'fs';
-import { resolve } from 'path';
 import { Plugin } from 'vite';
 import { run } from 'vite-plugin-run';
-
-/**
- * Options for configuring the Ziggy plugin.
- */
-export type Config = {
-  /**
-   * Path to the generated JavaScript/TypeScript file.
-   * @default 'node_modules/vite-plugin-ziggy/routes'
-   */
-  path?: string;
-
-  /**
-   * Whether to use `sail` instead of the `php` command.
-   * @default false
-   */
-  sail?: boolean;
-
-  /**
-   * Route group to generate.
-   */
-  group?: string;
-
-  /**
-   * Application URL.
-   */
-  url?: string;
-
-  /**
-   * Generate TypeScript declaration file.
-   * @default true
-   */
-  types?: boolean;
-
-  /**
-   * Generate only the TypeScript declaration file.
-   * @default true
-   */
-  typesOnly?: boolean;
-
-  /**
-   * Route name patterns to include.
-   * @default []
-   */
-  only?: string[];
-
-  /**
-   * Route name patterns to exclude.
-   * @default []
-   */
-  except?: string[];
-};
-
-const ZIGGY_PACKAGE_NAME = 'tightenco/ziggy';
-
-function getComposerPackageVersion(): string {
-  try {
-    const composerPath = resolve(process.cwd(), 'composer.json');
-    const composer = JSON.parse(readFileSync(composerPath, 'utf-8'));
-
-    if (!composer.require?.[ZIGGY_PACKAGE_NAME]) {
-      throw new Error(
-        `${ZIGGY_PACKAGE_NAME} not found in composer.json dependencies`,
-      );
-    }
-
-    const version = composer.require[ZIGGY_PACKAGE_NAME];
-    const match = version.match(/^[~^><]?(\d+)/);
-
-    if (!match) {
-      throw new Error(
-        `Invalid version format for ${ZIGGY_PACKAGE_NAME}: ${version}`,
-      );
-    }
-
-    const majorVersion = match[1];
-    return majorVersion;
-  } catch (error) {
-    if (
-      error instanceof Error &&
-      (error as NodeJS.ErrnoException).code === 'ENOENT'
-    ) {
-      throw new Error('composer.json not found in project root');
-    }
-    throw error;
-  }
-}
-
-function buildCommand(
-  version: string,
-  config: Required<Omit<Config, 'url' | 'group'>> & {
-    url?: string;
-    group?: string;
-  },
-): string[] {
-  const cmd = [
-    config.sail && !process.env.LARAVEL_SAIL ? 'sail' : 'php',
-    'artisan',
-    'ziggy:generate',
-    config.path,
-  ];
-
-  if (config.group) cmd.push('--group', config.group);
-  if (config.url) cmd.push('--url', config.url);
-
-  if (['1', '2'].includes(version)) {
-    if (config.types) cmd.push('--types');
-    if (config.typesOnly) cmd.push('--types-only');
-
-    if (version === '2') {
-      if (config.only.length > 0)
-        cmd.push('--only', config.only.join(','));
-      if (config.except.length > 0)
-        cmd.push('--except', config.except.join(','));
-    }
-  }
-
-  return cmd;
-}
+import { defaultConfig, Config } from './config.js';
+import { getComposerPackageVersion } from './utils.js';
+import { build, BuildConfig } from './build.js';
 
 export default (config: Config = {}): Plugin => {
-  const defaultConfig = {
-    path: 'node_modules/vite-plugin-ziggy/routess',
-    sail: false,
-    types: true,
-    typesOnly: true,
-    only: [],
-    except: [],
-  };
-
   try {
     const version = getComposerPackageVersion();
-    const cmd = buildCommand(version, { ...defaultConfig, ...config });
+    const cmd = build(version, { ...defaultConfig, ...config } as BuildConfig);
 
     const { configResolved, handleHotUpdate } = run([
       {
         name: 'ziggy-generator',
         run: cmd,
-        condition: (file) =>
-          file.includes('/routes/') && file.endsWith('.php'),
-        
+        condition: (file) => file.includes('/routes/') && file.endsWith('.php'),
       },
     ]);
     return {
@@ -162,5 +34,4 @@ export default (config: Config = {}): Plugin => {
   }
 
   return {} as Plugin;
-
 };
